requirements.txt
# Database
psycopg2-binary

# Environment
python-dotenv

# CLI Display
tabulate
colorama

# Email & Security
email-validator
bcrypt

# Development (optional)
pytest
pytest-cov